+++ src/magic.rs
@@ -19,10 +19,6 @@ use std::io::prelude::*;
 use std::io::{Error, ErrorKind};
 use std::process::*;
 
-/// The value of MAGIC_PARAM_BYTES_MAX borrowed from libmagic.
-/// See the man-page libmagic(3).
-const MAGIC_PARAM_BYTES_MAX: usize = 1_048_576;
-
 /// Whether the given MIME type denotes an SVG image.
 pub fn is_svg(mime: &Mime) -> bool {
     mime.type_() == mime::IMAGE && mime.subtype().as_str() == "svg"
@@ -43,13 +39,11 @@ pub fn detect_mime_type(buffer: &[u8]) -> Result<Mime, Box<dyn std::error::Error
         .stdin
         .as_mut()
         .expect("Forgot to pipe stdin?")
-        // extract only the first 1mb of interesting bytes; otherwise the
-        // method `write_all` fails with the "Broken pipe (os error 32)"
-        // error if the data being piped is large enough to be terminated
-        // due to signal SIGPIPE as the command file reads maximum
-        // 1048576 bytes from the input data.
-        // Check for MAGIC_PARAM_BYTES_MAX in the man-page of libmagic(3).
-        .write_all(&buffer[..std::cmp::min(buffer.len(), MAGIC_PARAM_BYTES_MAX)])?;
+        .write_all(buffer)
+        .or_else(|error| match error.kind() {
+            ErrorKind::BrokenPipe => Ok(()),
+            _ => Err(error),
+        })?;
 
     let output = process.wait_with_output()?;
     if output.status.success() {
@@ -96,7 +90,7 @@ mod tests {
     #[test]
     fn detect_mimetype_of_magic_param_bytes_max_length() {
         let data = std::iter::repeat(b'\0')
-            .take(MAGIC_PARAM_BYTES_MAX)
+            .take(1_048_576)
             .collect::<Vec<u8>>();
         let result = detect_mime_type(&data);
         assert!(result.is_ok(), "Unexpected error: {:?}", result);
@@ -105,7 +99,7 @@ mod tests {
     #[test]
     fn detect_mimetype_of_larger_than_magic_param_bytes_max_length() {
         let data = std::iter::repeat(b'\0')
-            .take(MAGIC_PARAM_BYTES_MAX * 2)
+            .take(1_048_576 * 2)
             .collect::<Vec<u8>>();
         let result = detect_mime_type(&data);
         assert!(result.is_ok(), "Unexpected error: {:?}", result);